Integrate latest changes from NBIS implementation to Biometric Evaluation framework using NBIS2EVAL.
Removes several warnings and adds support for AN2K Type-1 DCS fields indicating 7-bit ASCII.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 7776]

+++ nbis/lib/an2k/date.c
@@ -156,7 +156,8 @@ int get_ANSI_NIST_date(char **odate_str)
    /* Allocate ANSI/NIST date string "CCYYMMDD". */
    if((date_str = (char *)calloc(9, sizeof(char))) == NULL){
       fprintf(stderr, "ERROR : get_ANSI_NIST_date : "
-	      "calloc : date_str (%u bytes)\n", 9 * sizeof(char));
+	      "calloc : date_str (%lu bytes)\n",
+	      (unsigned long)(9 * sizeof(char)));
       return(-7);
    }
